requirements.txt
attrs==21.2.0
bip32utils==0.3.post4
certifi==2020.12.5
chardet==4.0.0
colored==1.4.2
discord-components==1.1.4
discord.py==1.7.3
ecdsa==0.16.1
idna==2.10
idna-ssl==1.1.0
mnemonic==0.19
multidict==5.1.0
Pillow==8.3.1
pip==21.1.3
PySocks==1.7.1
requests==2.25.1
setuptools==28.8.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.4
vk-api==11.9.4
yarl==1.6.3
youtube-dl==2021.3.14
